+++ umbrella/models/mistral.py
@@ -15,7 +15,7 @@ class Mistral(LLMBase):
         batch_size: int=1, 
         max_length: int=256, 
         device: str = "cuda:0", 
-        dtype = torch.float16) -> None:
+        dtype = torch.bfloat16) -> None:
 
         super().__init__()
         self.batch_size = batch_size
@@ -142,7 +142,7 @@ class Mistral(LLMBase):
 
 
 class MistralOffload(Mistral):
-    def __init__(self, model_name, batch_size = 1, max_length = 256, device = 'cuda:0', dtype=torch.float16):
+    def __init__(self, model_name, batch_size = 1, max_length = 256, device = 'cuda:0', dtype=torch.bfloat16):
         super().__init__(model_name, batch_size, max_length, device, dtype)
         self.load_stream = torch.cuda.Stream(device=device)
     
@@ -408,7 +408,7 @@ class MistralAwqOffload(MistralOffload):
 
 
 class MistralCudagraph(Mistral):
-    def __init__(self, model_name, batch_size = 1, max_length = 256, device = 'cuda:0', dtype=torch.float16):
+    def __init__(self, model_name, batch_size = 1, max_length = 256, device = 'cuda:0', dtype=torch.bfloat16):
         super().__init__(model_name, batch_size, max_length, device, dtype)
     
         self.callables = {}
